add verification

+++ src/modules/controllers/task.controller.js
@@ -8,28 +8,41 @@ module.exports.getAllTasks = (req, res, next) => {
 
 module.exports.createTask = (req, res, next) => {
   const task = new Task(req.body);
-  task.save().then(() => {
-    Task.find().then(result => {
-      res.send({ data: result });
-    });
-  }).catch(err => console.log(err));
+  const { body } = req;
+  if (body.hasOwnProperty('text')) {
+    task.save().then(() => {
+      Task.find().then(result => {
+        res.send({ data: result });
+      });
+    }).catch(err => console.log(err));
+  } else {
+    res.status(422).send('Error! Params not correct');
+  };
 };
 
 module.exports.changeTaskInfo = (req, res, next) => {
-  const reqObject = req.body;
-  const { id } = reqObject;
-  Task.updateOne({ _id: id }, reqObject).then(result => {
-    Task.find().then(result => {
-      res.send({ data: result });
+  const { body } = req;
+  const { _id } = body;
+  if (body.hasOwnProperty('_id') && (body.hasOwnProperty('text') || body.hasOwnProperty('isCheck'))) {
+    Task.updateOne({ _id: _id }, body).then(() => {
+      Task.find().then(result => {
+        res.send({ data: result });
+      });
     });
-  });
+  } else {
+    res.status(422).send('Error! Params not correct');
+  };
 };
 
 module.exports.deleteTask = (req, res) => {
-  const id = req.query.id;
-  Task.deleteOne({ _id: id }).then(() => {
-    Task.find().then(result => {
-      res.send({ data: result })
+  const { id } = req.query;
+  if (id) {
+    Task.deleteOne({ _id: id }).then(() => {
+      Task.find().then(result => {
+        res.send({ data: result })
+      });
     });
-  });
+  } else {
+    res.status(422).send('Error! Params not correct');
+  };
 };
